LPD-100614 Extract the individuals data set so both account pages can use it
The account individuals table was a data set hand-built inside the card that wraps it, which leaves nowhere for a second page to reuse it from. The columns, cell renderers, default sort and endpoint now live in an IndividualsDataSet component, following the AccountsDataSet pattern of a data set that owns its own definitions and takes what varies from its caller. AccountIndividuals keeps only its card shell. No rendered output changes.
